feat(cli): :sparkles: initial create command
Adds the initial command to create and bootstrap a new coat project.
`coat create <template>` will ask for a project name and create a
coat.json and package.json file in a new directory. It will also
install the template package and its peerDependencies and run the
setup and sync commands (not yet implemented).

See #4