feat(cli): :sparkles: Finish inital coat create command
Resolves the remaining todos for the coat create command and adds unit and integration tests.

I switched the template resolution to let it be fully handled by npm to allow the installation of local packages, tarballs, GitHub repositories and other sources that are supported by npm directly.

There are two types of tests: unit and integration tests.
Unit tests are meant to provide fast development feedback and test a single file while the integration tests run the cli in an isolated environment which mirrors how a user would install and use the cli.

Resolves #4