ci(cli): :construction_worker: Add GitHub Actions build and test workflow
Adds a workflow to lint, build and test the cli package.

The Lint, Test & Build job verifies that there are no linting issues, runs all unit tests and packs the cli package and uploads it as a build artifact.
Integration tests are run on both linux and mac, each with the Node.js LTS versions 10 and 12, as well as the current version 14.

Windows integration tests are currently not working correctly, since the tests time out as soon as they involve installing a template from a remote location. I plan on adding Windows integration tests as soon as possible and will try to locally get the tests running on a Windows machine.

Resolves #10