ci(cli): :construction_worker: add windows integration tests
Adds Windows to the test matrix for @coat/cli integration tests.

The integration tests were previously failing because the @coat/integration-test-template relied on a version of @coat/cli which had no node shebang at the top of the cli binary. In addition, the integration tests are no longer run concurrently since the parallel npm install commands sometimes lead to timeouts which resulted in flaky tests. The overall test timeout for integration tests has also been increased to 60 seconds, to give the tests enough time to perform an install of dependencies on slower network connections or systems.