feat(cli): :sparkles: Initial version of coat sync
Adds the `coat sync` cli command.

`coat sync` is the heart of the coat cli and responsible for generating
and keeping all project files up to date. The command can be executed in
a coat project to gather all files that are declared in the extended
coat templates, merge them and place them in the project directory. In
addition, it also merges declared scripts and dependencies into the
package.json file of the coat project and installs dependencies if they
have changed.

Resolves #5