ci(cli): :green_heart: Decrease macOS integration tests npm retry factors (#30)
macOS integration tests are still flaky, a reduction in the npm retry factor appears to help to get npm to retry network requests more frequently. 

It still appears that the GitHub Actions mac environment is very inconsistent when it comes to network performance. I'm not sure whether this solves the issue permanently.