feat(cli): :sparkles: Add coat lockfile
See # 31
Adds a lockfile that is generated by the `coat sync` cli command. The lockfile is in the yaml format and stored as `coat.lock` next to the coat manifest file.

It's purpose is to hold information that has been gathered from previous runs of `coat sync`. In the first version, it will only hold files that have been generated, in order to delete them after they are no longer generated by `coat sync`.