feat(cli): :sparkles: Pass merge function to custom file content functions
CoatManifestFiles can have a function for the content property, which allows them to perform advanced operations and merge or change the resulting file content.

In some scenarios however, it might be beneficial to have access to the default merge function that is used by coat, to not worry about the merging within the function and use the same merge logic that is used by coat.