feat(cli): :sparkles: Swap coat create arguments
Swaps the optional arguments that can be provided to coat create. The new syntax is:

`coat create <template> [directory] [projectName]`

The arguments have been swapped to make it easier to add coat to existing projects, by running `coat create template .`. Previously a project name would need to have been specified, for this to work as expected.

This still allows for the most typicial use case of calling `coat create template my-app-name`, which creates a directory "my-app-name" and initializes both the coat and package.json manifest files with "my-app-name" as the project name.

The created project folder is no longer deleted if there is an error during installation, e.g. when the template is unavailable. This behavior is changed to have a more predictable `coat create` command that never deletes creates files. Before this change coat create would have deleted created files only if it has been run in a non-existant or empty directory.