feat(cli): :sparkles: Make package.json optional for sync & setup
Adjusts the coat context type and mark the packageJson property as undefined if it does not exist in a project.

While this case should be rare, it allows to use coat projects that are exclusively using local templates and don't use any package.json specific properties like dependencies or scripts.