feat(cli): :sparkles: Allow creating coat projects in existing folders with files
Adjusts `coat create` to no longer check whether the target directory is empty before creating the project.

This change is made to make it easier to create a coat project in an already existing repository and start managing files via the specified template.

Previously it was not allowed to create a project in a directory that already contained files, since there were no safe guards to prevent accidental file loss, but since a recent commit added prompts before overwriting any files, it is now safe to create coat projects in existing projects.