fix(cli): :bug: Don't place once files if they exist already
There was a bug where once files would be placed if they already exist in the project before syncing, but did not yet exist in the lockfile.

This was because the check for the current file content used the relative path instead of the absolute file path in `src/sync/index.ts`.