feat(cli): :sparkles: Allow passing configuration to templates
Adds the possibility to pass a configuration object to extended coat templates.

`extends` entries in a coat manifest file or template can now also be tuples in addition to plain strings. E.g. the following usage calls the `my-template` template function with `{ "configValue": true }` true.

`"extends": [["my-template", { "configValue": true }]]`

If the template is a plain object, the provided config is ignored.