feat(cli): :sparkles: Remove existing scripts that conflict with coat managed scripts
`coat sync` now removes scripts that start with the same prefix that is used by a parallel merged script that has been added by coat.

This is done in order to avoid any unwanted behavior when adding coat or a new template to a project and then running additional scripts with the merged script in parallel, because it starts with the same prefix that is picked up from coat.