feat(template-ts-package): :sparkles: Add compiler config option
Adds a `compiler` configuration field that can be set to either `babel` or `typescript`. The template uses `typescript` by default if no option is specified.

If `typescript` is used, the build files are generated using the typescript compiler and jest is using `ts-jest` to run tests.

If `babel` is used, the build files are generated via babel and jest is using `babel-jest` to run tests.