fix(cli): :bug: Add space to recycling emoji for iTerm
iTerm - a popular macOS emulator emulator - has an issue with rendering certain emojis without any width. This issue is tracked in the iTerm repository: https://gitlab.com/gnachman/iterm2/-/issues/8735

In order to display the recycling emoji that is part of the "Everything up to date" message of `coat sync`, an additional space is added if the detected terminal app is iTerm.