chore(cli): :wrench: Remove bundle dependencies
Removes the `bundledDependencies` property from the package.json file, since it creates issues when installing dependencies both with npm v6 and v7. I though bundled dependencies could be used for packages that exist on the npm registry as well as packages from other sources, but it appears that npm packages were never supposed to be supported and only worked due to a bug in npm.