refactor(cli): :recycle: Switch create method arguments to options object
Changes the `create` method argument from multiple strings to an object that receives the parameters as properties.

This change is meant to improve the maintainability of the function when new arguments should be added to the create function without breaking changes.

BREAKING CHANGE: `create(template, directory, projectName)` is changed to `create({ template: "myTemplate", directory: ".", projectName: "myProject" })`