refactor(cli): :recycle: Switch setup method arguments to options object
Changes the `setup` method argument to an object that receives the current working directory and force flag in a property.

This change is meant to improve the maintainability of the function when new arguments should be added to the setup function without breaking changes.

BREAKING CHANGE: `setup(myCwd, true)` is changed to `setup({ cwd: myCwd, force: true })`