refactor(cli): :recycle: Switch run method arguments to options object
Changes the `run` method argument to an object that receives the parameters in a property.

This change is meant to improve the maintainability of the function when new arguments should be added to the run function without breaking changes.

BREAKING CHANGE: `run(cwd, scriptPatterns)` is changed to `run({ cwd: myCwd, scriptPatterns: myScripts })`