refactor(cli): :recycle: Switch sync method arguments to options object
Changes the `sync` method argument from a string to an object that receives the current working directory in a property.

This change is meant to improve the maintainability of the function when new arguments should be added to the sync function without breaking changes.

BREAKING CHANGE: `sync(myCwd)` is changed to `sync({ cwd: myCwd })`