feat(cli): :sparkles: Add sync --check option
`coat sync` now takes a `--check` option that checks whether all files that are generated by coat are up-to-date and exits without writing any change to the disk - even if there are pending changes.

If there are pending changes `coat` exits with the exit code 1.

This option is primarily targeted toward CI environments, to ensure that the coat project is up-to-date after changes have been made to its files or dependencies.

Fixes [CLI] Sync - Add --check flag #113