feat(cli): :sparkles: Run local cli if installed in a project
Local `@coat/cli` versions that are installed in projects should be preferred over a global installed version when coat is run directly from a command line.

This is done in order to support scenarios where a template depends on an older cli version and the globally installed version contains a breaking change.

Fixes #14 [CLI] Run local version of coat cli inside coat projects